+++ front/src/data/helpers/formatSearchParamToPrismaQuery.ts
